Plot Historic Proportions of Pulmonary/Extra-Pulmonary TB

Plot Historic Proportions of Pulmonary/Extra-Pulmonary TB

plot_historic_prop_pul_tb(df = tb_not_ew,
  plot_theme = ggplot2::theme_minimal,
  colour_scale = ggplot2::scale_fill_viridis_d, return = FALSE)

Arguments

df
A dataframe, (defaults to using tb_not_ew) of historic TB notifications

plot_theme
The ggplot2 theme to use, defaults to theme_minimal.

colour_scale
The colour scale to plot with, defaults to scale_colour_viridis_d

return
A logical indiciting if the plot should be printed or returned (defaults to FALSE)

Value

A ggplot2 plot of TB notifications over time, with a secondary zoomed plot from a specified date.

Examples

plot_historic_prop_pul_tb(plot_theme = ggplot2::theme_minimal, colour_scale = ggplot2::scale_fill_viridis_d)

#> Error: 'scale_fill_viridis_d' is not an exported object from 'namespace:ggplot2'